+++ tests/testthat/test_parse.R
@@ -4,6 +4,5 @@ context("Parsing function accuracy")
 test_that("Error codes", {
   resp <- httr:::response()
   resp$status_code <- 201
- expect_warning(parse_response(resp))
-
+ expect_error(parse_response(resp))
 })